requirements.txt
streamlit
selenium
telegram
python-telegram-bot
Pillow
pytesseract
cryptography
webdriver-manager
